Modified BaseDAO and AbstractModel from Long to int!

+++ src/main/java/ro/sci/digitalCookBook/dao/BaseDAO.java
@@ -7,7 +7,7 @@ public interface BaseDAO<T extends AbstractModel> {
 
     Collection<T> getAll();
 
-    T findById(Long id);
+    T findById(int id);
 
     T update(T model);
